[#16] fix: 암장 검색 레이아웃 하나의 파일에서 관리하도록 수정
- 암장 선택 이벤트 핸들러 SearchPlace 컴포넌트 prop에 추가
