--- src/app/place/page.tsx
@@ -1,5 +1,0 @@
-import Place from "@/app/place/Place";
-
-export default function Page() {
-  return <Place />;
-}
